+++ docs/api/index.rst
@@ -27,7 +27,7 @@ API Reference
       :link: metrics
       :link-type: doc
 
-      Built-in metric presets and the ``MetricFn`` callable for ``summary(metrics=...)``.
+      Built-in metric presets and the metric types for ``summary(metrics=...)``.
 
    .. grid-item-card:: Testing
       :link: testing
